requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
load-dotenv==0.1.0
mailchimp3==3.0.21
psycopg2==2.9.9
pygame==2.5.2
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
